Start moving the `emit_trace` logic into `parse_to_trace`.
The idea is to move towards emitting directly emit_dtos from parse_dtos.
