+++ text_to_perfetto.py
@@ -1,10 +1,9 @@
 #!env python3
 
 import argparse
-from lib.emit_trace import emit_trace
 from lib.perfetto_writer import PerfettoWriter
 from lib.parse_text_trace import parse_text_trace
-from lib.parse_to_trace import parse_to_trace
+from lib.parse_to_trace import parse_to_trace, emit_trace
 
 
 def main():
